% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{print.ipss_result}
\alias{print.ipss_result}
\title{Print method for ipss_result}
\usage{
\method{print}{ipss_result}(x, ...)
}
\arguments{
\item{x}{An \code{ipss_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for ipss_result
}
