% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability_selection.R
\name{summary.ipss_result}
\alias{summary.ipss_result}
\title{Summary method for ipss_result}
\usage{
\method{summary}{ipss_result}(object, ...)
}
\arguments{
\item{object}{An \code{ipss_result} object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for ipss_result
}
